requirements.txt
openai~=1.55.2
pdf2image~=1.17.0
python-dotenv~=1.0.1